apps/appointment-service/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
motor==3.3.2
pymongo==4.5.0
PyJWT==2.8.0
pydantic==2.4.2
